Fix #381
Formating obs Date and Datetime values to prevent angular $httpParamSerializer from converting them
to an ISO 8601 string before submitting.

Angular converts Date objects to ISO 8601 Strings with timezones and in OpenMRS Obs#setValueAsString
the timezone is not considered when parsing to Date or Datetime. In some cases this caused the wrong
dates to be saved in the database.